3 Libraries and Frameworks/9 REST/example2/requirements.txt
fastapi
uvicorn
pyjwt
pydantic
slowapi
